requirements.txt
anyio==4.12.1
certifi==2026.1.4
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
psycopg2-binary==2.9.11
python-decouple==3.8
python-telegram-bot==22.6
SQLAlchemy==2.0.46
typing_extensions==4.15.0
